% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataRobot_cols}
\alias{DataRobot_cols}
\title{Function to extract DataRobot colors as hex codes}
\usage{
DataRobot_cols(...)
}
\arguments{
\item{...}{Character names of DataRobot colors}
}
\description{
Function to extract DataRobot colors as hex codes
}
