% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataRobot_get_palettes}
\alias{DataRobot_get_palettes}
\title{Return function to return a list of DataRobot color palettes}
\usage{
DataRobot_get_palettes(palette = palette)
}
\description{
Return function to return a list of DataRobot color palettes
}
