% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataRobot_pal}
\alias{DataRobot_pal}
\title{Return function to interpolate a DataRobot color palette}
\usage{
DataRobot_pal(palette = "DR_Blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in DataRobot_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a DataRobot color palette
}
