% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataRobot_palette_names}
\alias{DataRobot_palette_names}
\title{Return the names of DataRobot color palettes}
\usage{
DataRobot_palette_names()
}
\description{
Return the names of DataRobot color palettes
}
