% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_color_DataRobot}
\alias{scale_color_DataRobot}
\title{Color scale constructor for DataRobot colors}
\usage{
scale_color_DataRobot(
  palette = "DR_Blues",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in DataRobot_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for DataRobot colors
}
